requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1

google-cloud-storage==2.18.2

torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.8
pillow>=10.3.0
numpy>=1.26.4


python-multipart==0.0.9

paho-mqtt

plotly
streamlit
